perf(fmm): run real-basis rotation GEMMs in the working dtype (fp32), not fp64
compute_real_B_matrix_{multipole,local} took a `dtype` argument but used it only
as the lru_cache key (`str(dtype)`) and always returned the float64-computed B
matrix via `jnp.asarray(B_U)`. Under `JAX_ENABLE_X64=1` that left the real-basis
rotate-to-z / rotate-from-z blocks in float64, so every M2M, L2L and M2L
rotation-scale GEMM ran as a float64 cuBLAS/cutlass GEMM (`f64 block @ f32 coeffs
-> f64`) even though the multipole/local coefficients are float32. On the A100
large-N fused fast lane those fp64 GEMMs were ~15% of per-step device time.

Cast the (still float64-computed, for accuracy) B matrix down to the requested
dtype at the function boundary so the downstream rotation GEMMs run in the working
dtype. The z-translate tables already honored dtype; this was the only leak.

A100, 200k / order 4 / theta 0.8 / leaf 256, real-basis device-only fused fast
lane (Pallas near-field + M2L, treecode walk):

  - 91.3 -> 81.2 ms/step (-11%, 1.12x)
  - force parity vs direct unchanged: rel-err p50 1.161e-3 -> 1.158e-3 (p90/max
    identical)
  - energy/Lz conservation identical to fp64 (A/B: equilibrium-IC max|dE/E0|
    0.0878249 vs 0.0878251; generate-IC 0.10166 vs 0.10167)
  - fast lane stays engaged (fastlane_hits=1, fallback_count=0)
  - perfetto: fp64 d884 GEMMs (~15%) replaced by fp32 gemm_fusion (~3.6%),
    device time 90 -> 79.5 ms/step

tests/unit/operators/test_real_harmonics.py: 39/39 pass.

+++ jaccpot/operators/real_harmonics.py
@@ -1095,7 +1095,11 @@ def compute_real_B_matrix_local(ell: int, *, dtype: DTypeLike) -> Array:
         Real B_T matrix of shape (2*ell+1, 2*ell+1).
     """
     B_T, _ = _compute_B_real_dehnen_via_Q(ell, str(dtype))
-    return jnp.asarray(B_T)
+    # Honor the requested dtype: the B matrix is computed in float64 for accuracy
+    # but must be cast down to the working dtype so the downstream rotation GEMMs
+    # (M2M/L2L/M2L rotate-to-z/from-z) run in that dtype rather than promoting the
+    # float32 coefficient vectors to float64 (the fp64 cuBLAS GEMM slowdown).
+    return jnp.asarray(B_T, dtype=dtype)
 
 
 def compute_real_B_matrix_multipole(ell: int, *, dtype: DTypeLike) -> Array:
@@ -1119,7 +1123,10 @@ def compute_real_B_matrix_multipole(ell: int, *, dtype: DTypeLike) -> Array:
         Real B_U matrix of shape (2*ell+1, 2*ell+1).
     """
     _, B_U = _compute_B_real_dehnen_via_Q(ell, str(dtype))
-    return jnp.asarray(B_U)
+    # Honor the requested dtype (see compute_real_B_matrix_local): cast the
+    # float64-computed B matrix down to the working dtype so the rotation GEMMs
+    # do not run in float64.
+    return jnp.asarray(B_U, dtype=dtype)
 
 
 def verify_real_B_matrix(ell: int, *, dtype: DTypeLike) -> Tuple[bool, float, float]:
